STYLE: NAV-76 - Avoid nested parallel streams in transfer generators
Bad practice to nest parallel streams and since the same station transfer is really efficient (maybe 100ms), it is not worth to run this in parallel. If we have some really demanding generators in future, we can again think about parallel execution.

But for now the important loop to be executed concurrently is inside the WalkTransferGenerator.